requirements.txt
flask
tensorflow
Pillow<=9.5.0
pandas
matplotlib
scikit-learn
numpy
keras_cv